changing paquetes & prendas structure

+++ src/app/components/house-globe/house-globe.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { GdevCache } from '@jgu7man/gdev-tools';
+import { GdevAlert, GdevCache, GdevMessageAlertModel } from '@jgu7man/gdev-tools';
 import { Observable, of } from 'rxjs';
 import { map, tap } from 'rxjs/operators';
 
@@ -14,7 +14,8 @@ export class HouseGlobeComponent implements OnInit {
   propOpened: Observable<any>
   constructor(
     private _cache: GdevCache,
-    private _router: Router
+    private _router: Router,
+    private _alert: GdevAlert
   ) {
     this.propOpened = this._cache
       .listenForChanges('currentProp')
@@ -31,10 +32,17 @@ export class HouseGlobeComponent implements OnInit {
   }
 
   closeProp() {
-    let perfil = window.location.href.split('/')[3]
-    this._cache.deleteDataKey('currentProp')
-    this._cache.deleteDataKey('prendasReport')
-    this._router.navigate([`/${perfil}`])
+    let bodyAlert: GdevMessageAlertModel = {
+      message: 'No has terminado de registrar, ¿Seguro que quieres cerrar la propiedad?', trueMsg: 'Sí', falseMsg: 'No'
+    }
+    this._alert.sendRequestAlert(bodyAlert).subscribe(confirm => {
+      if (confirm) {
+        let perfil = window.location.href.split('/')[3]
+        this._cache.deleteDataKey('currentProp')
+        this._cache.deleteDataKey('prendasReport')
+        this._router.navigate([`/${perfil}`])
+      }
+    })
   }
 
   seeProp() {
